Move all /api/comment routes to its own file
